Move RegionAnnotation to annotation.py
