+++ claripy/vsa/__init__.py
@@ -1,4 +1,4 @@
-from .valueset import RegionAnnotation, ValueSet
+from .valueset import ValueSet
 from .strided_interval import StridedInterval, CreateStridedInterval
 from .discrete_strided_interval_set import DiscreteStridedIntervalSet
 from .bool_result import BoolResult, TrueResult, FalseResult, MaybeResult
